Create separate sub pages for home
