+++ src/pages/Home/Landing.js
@@ -1,0 +1,16 @@
+import React from 'react';
+import Container from 'react-bootstrap/Container';
+
+/**
+ * Landing page for Pan-Lang.
+ * The first thing people will see when visiting our web app.
+ */
+function Landing() {
+  return (
+    <Container>
+      <p>We'll make this pretty later, just sign in.</p>
+    </Container>
+  );
+}
+
+export default Landing;
